refactor: remove unused imports and variables and checked exceptions
