+++ src/test/java/woowacourse/crewniverse/milkyway/service/response/AttendanceSheetResponseTest.java
@@ -2,7 +2,6 @@ package woowacourse.crewniverse.milkyway.service.response;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import woowacourse.crewniverse.milkyway.service.response.AttendanceSheetResponse;
 
 class AttendanceSheetResponseTest {
